OPT_TYPE

+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/ExchComp.cxx
@@ -185,7 +185,7 @@ cxxExchComp::read_raw(CParser & parser, bool check)
 	std::string token;
 	int opt_save;
 
-	opt_save = CParser::OPT_TYPE::OPT_ERROR;
+	opt_save = (int)CParser::OPT_TYPE::OPT_ERROR;
 	bool la_defined(false);
 	bool charge_balance_defined(false);
 	bool formula_z_defined(false);
@@ -193,20 +193,20 @@ cxxExchComp::read_raw(CParser & parser, bool check)
 	for (;;)
 	{
 		int opt = parser.get_option(vopts, next_char);
-		if (opt == CParser::OPT_TYPE::OPT_DEFAULT)
+		if (opt == (int)CParser::OPT_TYPE::OPT_DEFAULT)
 		{
 			opt = opt_save;
 		}
 
 		switch (opt)
 		{
-		case CParser::OPT_TYPE::OPT_EOF:
+		case (int)CParser::OPT_TYPE::OPT_EOF:
 			break;
-		case CParser::OPT_TYPE::OPT_KEYWORD:
+		case (int)CParser::OPT_TYPE::OPT_KEYWORD:
 			break;
-		case CParser::OPT_TYPE::OPT_DEFAULT:
-		case CParser::OPT_TYPE::OPT_ERROR:
-			opt = CParser::OPT_TYPE::OPT_KEYWORD;
+		case (int)CParser::OPT_TYPE::OPT_DEFAULT:
+		case (int)CParser::OPT_TYPE::OPT_ERROR:
+			opt = (int)CParser::OPT_TYPE::OPT_KEYWORD;
 			// Allow return to Exchange for more processing
 			break;
 
@@ -307,7 +307,7 @@ cxxExchComp::read_raw(CParser & parser, bool check)
 			parser.warning_msg("-formula_totals is an obsolete identifier");
 			break;
 		}
-		if (opt == CParser::OPT_TYPE::OPT_EOF || opt == CParser::OPT_TYPE::OPT_KEYWORD)
+		if (opt == (int)CParser::OPT_TYPE::OPT_EOF || opt == (int)CParser::OPT_TYPE::OPT_KEYWORD)
 			break;
 	}
 	if (check)
