+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/PPassemblage.cxx
@@ -138,7 +138,7 @@ cxxPPassemblage::read_raw(CParser & parser, bool check)
 	this->read_number_description(parser);
 	this->Set_new_def(false);
 
-	opt_save = CParser::OPT_TYPE::OPT_ERROR;
+	opt_save = (int)CParser::OPT_TYPE::OPT_ERROR;
 
 	for (;;)
 	{
@@ -151,20 +151,20 @@ cxxPPassemblage::read_raw(CParser & parser, bool check)
 		{
 			opt = parser.getOptionFromLastLine(vopts, next_char, true);
 		}
-		if (opt == CParser::OPT_TYPE::OPT_DEFAULT)
+		if (opt == (int)CParser::OPT_TYPE::OPT_DEFAULT)
 		{
 			opt = opt_save;
 		}
 		useLastLine = false;
 		switch (opt)
 		{
-		case CParser::OPT_TYPE::OPT_EOF:
+		case (int)CParser::OPT_TYPE::OPT_EOF:
 			break;
-		case CParser::OPT_TYPE::OPT_KEYWORD:
+		case (int)CParser::OPT_TYPE::OPT_KEYWORD:
 			break;
-		case CParser::OPT_TYPE::OPT_DEFAULT:
-		case CParser::OPT_TYPE::OPT_ERROR:
-			opt = CParser::OPT_TYPE::OPT_EOF;
+		case (int)CParser::OPT_TYPE::OPT_DEFAULT:
+		case (int)CParser::OPT_TYPE::OPT_ERROR:
+			opt = (int)CParser::OPT_TYPE::OPT_EOF;
 			parser.
 				error_msg("Unknown input in EQUILIBRIUM_PHASES_RAW keyword.",
 						  PHRQ_io::OT_CONTINUE);
@@ -231,7 +231,7 @@ cxxPPassemblage::read_raw(CParser & parser, bool check)
 			opt_save = 3;
 			break;
 		}
-		if (opt == CParser::OPT_TYPE::OPT_EOF || opt == CParser::OPT_TYPE::OPT_KEYWORD)
+		if (opt == (int)CParser::OPT_TYPE::OPT_EOF || opt == (int)CParser::OPT_TYPE::OPT_KEYWORD)
 			break;
 	}
 }
