+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/PPassemblageComp.cxx
@@ -111,7 +111,7 @@ cxxPPassemblageComp::read_raw(CParser & parser, bool check)
 	std::string token;
 	int opt_save;
 
-	opt_save = CParser::OPT_TYPE::OPT_ERROR;
+	opt_save = (int)CParser::OPT_TYPE::OPT_ERROR;
 	bool si_defined(false);
 	bool moles_defined(false);
 	bool delta_defined(false);
@@ -122,20 +122,20 @@ cxxPPassemblageComp::read_raw(CParser & parser, bool check)
 	for (;;)
 	{
 		int opt = parser.get_option(vopts, next_char);
-		if (opt == CParser::OPT_TYPE::OPT_DEFAULT)
+		if (opt == (int)CParser::OPT_TYPE::OPT_DEFAULT)
 		{
 			opt = opt_save;
 		}
 
 		switch (opt)
 		{
-		case CParser::OPT_TYPE::OPT_EOF:
+		case (int)CParser::OPT_TYPE::OPT_EOF:
 			break;
-		case CParser::OPT_TYPE::OPT_KEYWORD:
+		case (int)CParser::OPT_TYPE::OPT_KEYWORD:
 			break;
-		case CParser::OPT_TYPE::OPT_DEFAULT:
-		case CParser::OPT_TYPE::OPT_ERROR:
-			opt = CParser::OPT_TYPE::OPT_KEYWORD;
+		case (int)CParser::OPT_TYPE::OPT_DEFAULT:
+		case (int)CParser::OPT_TYPE::OPT_ERROR:
+			opt = (int)CParser::OPT_TYPE::OPT_KEYWORD;
 			// Allow return to Exchange for more processing
 			break;
 
@@ -262,7 +262,7 @@ cxxPPassemblageComp::read_raw(CParser & parser, bool check)
 			opt_save = 10;
 			break;
 		}
-		if (opt == CParser::OPT_TYPE::OPT_EOF || opt == CParser::OPT_TYPE::OPT_KEYWORD)
+		if (opt == (int)CParser::OPT_TYPE::OPT_EOF || opt == (int)CParser::OPT_TYPE::OPT_KEYWORD)
 			break;
 	}
 	// members that must be defined
