+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/Pressure.cxx
@@ -210,7 +210,7 @@ cxxPressure::read_raw(CParser & parser, bool check)
 	// Number and description set in read_reaction_pressure_raw
 	this->read_number_description(parser);
 
-	opt_save = CParser::OPT_TYPE::OPT_ERROR;
+	opt_save = (int)CParser::OPT_TYPE::OPT_ERROR;
 	bool equalIncrements_defined(false);
 	bool count_defined(false);
 
@@ -225,19 +225,19 @@ cxxPressure::read_raw(CParser & parser, bool check)
 		{
 			opt = parser.getOptionFromLastLine(vopts, next_char, true);
 		}
-		if (opt == CParser::OPT_TYPE::OPT_DEFAULT)
+		if (opt == (int)CParser::OPT_TYPE::OPT_DEFAULT)
 		{
 			opt = opt_save;
 		}
 		switch (opt)
 		{
-		case CParser::OPT_TYPE::OPT_EOF:
+		case (int)CParser::OPT_TYPE::OPT_EOF:
 			break;
-		case CParser::OPT_TYPE::OPT_KEYWORD:
+		case (int)CParser::OPT_TYPE::OPT_KEYWORD:
 			break;
-		case CParser::OPT_TYPE::OPT_DEFAULT:
-		case CParser::OPT_TYPE::OPT_ERROR:
-			opt = CParser::OPT_TYPE::OPT_EOF;
+		case (int)CParser::OPT_TYPE::OPT_DEFAULT:
+		case (int)CParser::OPT_TYPE::OPT_ERROR:
+			opt = (int)CParser::OPT_TYPE::OPT_EOF;
 			parser.error_msg("Unknown input in REACTION_PRESSURE_RAW keyword.",
 							 PHRQ_io::OT_CONTINUE);
 			parser.error_msg(parser.line().c_str(), PHRQ_io::OT_CONTINUE);
@@ -275,7 +275,7 @@ cxxPressure::read_raw(CParser & parser, bool check)
 				parser.incr_input_error();
 				parser.error_msg("Expected boolean value for equalIncrements.", PHRQ_io::OT_CONTINUE);
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			useLastLine = false;
 			equalIncrements_defined = true;
 			break;
@@ -287,12 +287,12 @@ cxxPressure::read_raw(CParser & parser, bool check)
 				parser.incr_input_error();
 				parser.error_msg("Expected integer value for count.", PHRQ_io::OT_CONTINUE);
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			useLastLine = false;
 			count_defined = true;
 			break;
 		}
-		if (opt == CParser::OPT_TYPE::OPT_EOF || opt == CParser::OPT_TYPE::OPT_KEYWORD)
+		if (opt == (int)CParser::OPT_TYPE::OPT_EOF || opt == (int)CParser::OPT_TYPE::OPT_KEYWORD)
 			break;
 	}
 	// members that must be defined
