+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/SS.cxx
@@ -131,7 +131,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 	std::string token;
 	int opt_save;
 
-	opt_save = CParser::OPT_TYPE::OPT_ERROR;
+	opt_save = (int)CParser::OPT_TYPE::OPT_ERROR;
 	bool a0_defined(false);
 	bool a1_defined(false);
 	bool ag0_defined(false);
@@ -157,26 +157,26 @@ cxxSS::read_raw(CParser & parser, bool check)
 		{
 			opt = parser.getOptionFromLastLine(vopts, next_char, false);
 		}
-		if (opt == CParser::OPT_TYPE::OPT_DEFAULT)
+		if (opt == (int)CParser::OPT_TYPE::OPT_DEFAULT)
 		{
 			opt = opt_save;
 		}
 		useLastLine = false;
 		switch (opt)
 		{
-		case CParser::OPT_TYPE::OPT_EOF:
+		case (int)CParser::OPT_TYPE::OPT_EOF:
 			break;
-		case CParser::OPT_TYPE::OPT_KEYWORD:
+		case (int)CParser::OPT_TYPE::OPT_KEYWORD:
 			break;
-		case CParser::OPT_TYPE::OPT_DEFAULT:
-		case CParser::OPT_TYPE::OPT_ERROR:
-			opt = CParser::OPT_TYPE::OPT_KEYWORD;
+		case (int)CParser::OPT_TYPE::OPT_DEFAULT:
+		case (int)CParser::OPT_TYPE::OPT_ERROR:
+			opt = (int)CParser::OPT_TYPE::OPT_KEYWORD;
 			// Allow return to Exchange for more processing
 			break;
 
 		case 0:				// name not used
 			parser.warning_msg("-ss_name not used, defined in -solid_solution");
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 1:				// total_moles
@@ -186,7 +186,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 				parser.incr_input_error();
 				parser.error_msg("Expected numeric value for total_moles.", PHRQ_io::OT_CONTINUE);
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 2:				// a0
@@ -198,7 +198,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			a0_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 3:				// a1
@@ -210,7 +210,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			a1_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 4:				// components
@@ -251,7 +251,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 				}
 			}
 			
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 5:				// miscibility
@@ -263,7 +263,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			miscibility_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 6:				// spinodal
@@ -273,7 +273,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 				parser.incr_input_error();
 				parser.error_msg("Expected boolean value for spinodal.", PHRQ_io::OT_CONTINUE);
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 7:				// tk
@@ -283,7 +283,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 				parser.incr_input_error();
 				parser.error_msg("Expected numeric value for tk.", PHRQ_io::OT_CONTINUE);
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 8:				// xb1
@@ -295,7 +295,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			xb1_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 9:				// xb2
@@ -307,7 +307,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			xb2_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 10:				// ag0
@@ -319,7 +319,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			ag0_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 
 		case 11:				// ag1
@@ -331,7 +331,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 								 PHRQ_io::OT_CONTINUE);
 			}
 			ag1_defined = true;
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 		case 13:				// input_case
 			{
@@ -347,7 +347,7 @@ cxxSS::read_raw(CParser & parser, bool check)
 					this->input_case = (cxxSS::SS_PARAMETER_TYPE) i;
 				}
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 		case 14:				// p
 			{
@@ -389,10 +389,10 @@ cxxSS::read_raw(CParser & parser, bool check)
 				parser.incr_input_error();
 				parser.error_msg("Expected numeric value for dn.", PHRQ_io::OT_CONTINUE);
 			}
-			opt_save = CParser::OPT_TYPE::OPT_DEFAULT;
+			opt_save = (int)CParser::OPT_TYPE::OPT_DEFAULT;
 			break;
 		}
-		if (opt == CParser::OPT_TYPE::OPT_EOF || opt == CParser::OPT_TYPE::OPT_KEYWORD)
+		if (opt == (int)CParser::OPT_TYPE::OPT_EOF || opt == (int)CParser::OPT_TYPE::OPT_KEYWORD)
 			break;
 	}
 	if (check)
